add test for changing github repository #39
downloading with get_doctops.sh

+++ tests/get_doctops.bats
@@ -6,7 +6,7 @@
 @test "download binary with get_docopts.sh" {
   # go to repository basepath
   cd ..
-  rm docopts
+  [[ -x docopts ]] && rm docopts
   run ./get_docopts.sh
   echo "$output"
   [[ $status -eq 0 ]]
@@ -20,5 +20,27 @@
   [[ $status -eq 0 ]]
   # test version match
   version=$(cat VERSION)
+  [[ $output =~ $version ]]
+}
+
+
+@test "get_docopts.sh use another repository" {
+  # go to repository basepath
+  cd ..
+  [[ -x docopts ]] && rm docopts
+  export GIT_USER=Sylvain303
+  export RELEASE=v0.6.4-alpha1
+  run ./get_docopts.sh
+  echo "$output"
+  [[ $status -eq 0 ]]
+
+  match_url='Fetching from: https://github.com/Sylvain303/docopts/'
+  [[ $output =~ $match_url ]]
+
+  [[ -x docopts ]]
+  run ./docopts --version
+  [[ $status -eq 0 ]]
+  # test version match
+  version='v0.6.4-alpha1'
   [[ $output =~ $version ]]
 }
